Refactor: Improved m24c32 driver error handling and partition access.

+++ general/include/m24c32.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 #include <string.h>
 
-typedef int (*write_ptr)(uint8_t addr, const uint8_t *data, uint8_t len);
-typedef int (*read_ptr)(uint8_t addr, uint8_t *data, uint8_t len);
+typedef int (*write_ptr)(uint16_t addr, const uint8_t *data, uint16_t len);
+typedef int (*read_ptr)(uint16_t addr, uint8_t *data, uint16_t len);
 
 typedef struct {
 	write_ptr write;
@@ -20,7 +20,7 @@ typedef struct {
  * @param addr Memory address of the write head.
  * @param data Buffer of bytes to write.
  * @param len The amount of bytes to write.
- * @return int Error code.
+ * @return int Error code: 0 on success, nonzero on failure.
  */
 int m24c32_write(m24c32_t *device, uint16_t addr, uint8_t *data, uint16_t len);
 
@@ -31,7 +31,7 @@ int m24c32_write(m24c32_t *device, uint16_t addr, uint8_t *data, uint16_t len);
  * @param addr Memory address of the read head.
  * @param data Buffer where data will be written to.
  * @param len The amount of bytes to read.
- * @return int Error code.
+ * @return int Error code: 0 on success, nonzero on failure.
  */
 int m24c32_read(m24c32_t *device, uint16_t addr, uint8_t *data, uint16_t len);
 
@@ -41,7 +41,7 @@ int m24c32_read(m24c32_t *device, uint16_t addr, uint8_t *data, uint16_t len);
  * @param device I2C read and write function pointers.
  * @param addr Memory address of the write head.
  * @param len Number of bytes to clear.
- * @return int Error code.
+ * @return int Error code: 0 on success, nonzero on failure.
  */
 int m24c32_clear(m24c32_t *device, uint16_t addr, uint16_t len);
